feat: move Placeholder to a component folder
