+++ src/components/Page/Page.jsx
@@ -6,30 +6,10 @@ import PageLinks from '../PageLinks/PageLinks';
 import Markdown from '../Markdown/Markdown';
 import Contributors from '../Contributors/Contributors';
 import Gitter from '../Gitter/Gitter';
+import Placeholder from '../Placeholder/Placeholder';
 
 // Load Styling
 import './Page.scss';
-
-// Placeholder string
-const Placeholder = () => (`
-  <div class="placeholder">
-    <p class="placeholder--medium">&nbsp;</p>
-    <p class="placeholder--large">&nbsp;</p>
-    <p class="placeholder--small">&nbsp;</p>
-    <h2 class="placeholder--xsmall">&nbsp;</h2>
-    <p class="placeholder--large">&nbsp;</p>
-    <p class="placeholder--small">&nbsp;</p>
-    <p class="placeholder--medium">&nbsp;</p>
-    <h2 class="placeholder--small">&nbsp;</h2>
-    <p class="placeholder--large">&nbsp;</p>
-    <p class="placeholder--medium">&nbsp;</p>
-    <p class="placeholder--xsmall">&nbsp;</p>
-    <h2 class="placeholder--xsmall">&nbsp;</h2>
-    <p class="placeholder--large">&nbsp;</p>
-    <p class="placeholder--small">&nbsp;</p>
-    <p class="placeholder--medium">&nbsp;</p>
-  </div>
-`);
 
 class Page extends React.Component {
   constructor(props) {
