+++ src/components/Placeholder/Placeholder.js
@@ -1,0 +1,25 @@
+// Import Styling
+import './Placeholder.scss';
+
+// Placeholder string
+const Placeholder = () => (`
+  <div class="placeholder">
+    <p class="placeholder--medium">&nbsp;</p>
+    <p class="placeholder--large">&nbsp;</p>
+    <p class="placeholder--small">&nbsp;</p>
+    <h2 class="placeholder--xsmall">&nbsp;</h2>
+    <p class="placeholder--large">&nbsp;</p>
+    <p class="placeholder--small">&nbsp;</p>
+    <p class="placeholder--medium">&nbsp;</p>
+    <h2 class="placeholder--small">&nbsp;</h2>
+    <p class="placeholder--large">&nbsp;</p>
+    <p class="placeholder--medium">&nbsp;</p>
+    <p class="placeholder--xsmall">&nbsp;</p>
+    <h2 class="placeholder--xsmall">&nbsp;</h2>
+    <p class="placeholder--large">&nbsp;</p>
+    <p class="placeholder--small">&nbsp;</p>
+    <p class="placeholder--medium">&nbsp;</p>
+  </div>
+`);
+
+export default Placeholder;
